ui: Remove unnecessay reopens from sort and form services
